76179: Move DateMathParser to dspace-api module
DateMathParser will be needed in AccessConditionOption, so this
change is needed to avoid a circular dependency
(dspace-api -> dspace-server-webapp).
